backend/requirements.txt
aiohttp
python-socketio
websockets
python-telegram-bot
python-dotenv
requests

pip install aiohttp aiohttp-cors python-socketio websockets python-telegram-bot python-dotenv requests